requirements.txt
pygame==1.9.4
RPi.GPIO==0.6.4
